langchain_playground/Tools/YouTubeLoader/requirements.txt
fal-client
replicate
ipython
more-itertools
numpy
openai-whisper
opencc-python-reimplemented
optimum
pytubefix
python-dotenv
tiktoken
torch
transformers
tqdm